commande fonctionnelle avant rework
