+++ src/main/resources/plugin.yml
@@ -142,5 +142,5 @@ commands:
     description: Permet d'utiliser un préfixe par défaut. Utilisation /prefix <prefixe>
   TechniqueMJ:
     description: Permet d'utiliser une technique MJ
-  Prier:
-    description: Permet de prier le Rikudo
+  Priere:
+    description: Permet d'écrire une prière le Rikudo
